% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calConnect.R
\name{calConnect}
\alias{calConnect}
\title{Connect to calibrations database}
\usage{
calConnect(path = "default", timeout = 1e+05, silent = FALSE)
}
\arguments{
\item{path}{The path to the database. 'default' points to "//env-fs/env-data/corp/water/Common_GW_SW/Data/calibrations/WRBcalibrates.sqlite".}

\item{timeout}{The duration in which to retry an operation in milliseconds. Valid for the duration of the connection.}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function exists to facilitate connecting to the calibrations database. Cannot be used to create a new database, use \code{\link[DBI:dbConnect]{DBI::dbConnect()}} with the appropriate driver (e.g. \code{\link[RSQLite:SQLite]{RSQLite::SQLite()}}) for that purpose.
}
