% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{db_create}
\alias{db_create}
\title{Initial calibration database creation}
\usage{
db_create(con = calConnect(), overwrite = FALSE)
}
\arguments{
\item{con}{A connection to the target database}

\item{overwrite}{TRUE overwrites the database, if one exists in the same path.}
}
\value{
New tables in the target postgres database.
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Populates a postgres database with the necessary tables for the calibration app. A database should be created before using this function, but it can be left empty.

This function will also establish a mapping to the 'hydromet' database, which is used to populate the 'locations' table.
}
