% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRBcalibrates-package.R
\docType{package}
\name{WRBcalibrates-package}
\alias{WRBcalibrates}
\alias{WRBcalibrates-package}
\title{WRBcalibrates: WRB calibration app}
\description{
App to save instrument calibration data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YukonWRB/WRBcalibrates}
  \item Report bugs at \url{https://github.com/YukonWRB/WRBcalibrates/issues}
}

}
\author{
\strong{Maintainer}: Ghislain de Laplante \email{ghislain.delaplante@yukon.ca}

}
\keyword{internal}
