% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calConnect.R
\name{calConnect}
\alias{calConnect}
\title{Connect to the calibrations database}
\usage{
calConnect(
  name = "calibrations",
  host = Sys.getenv("calHost"),
  port = Sys.getenv("calPort"),
  username = Sys.getenv("calUser"),
  password = Sys.getenv("calPass"),
  silent = FALSE
)
}
\arguments{
\item{name}{Database name.}

\item{host}{Database host address. By default searches the .Renviron file for parameter:value pair of form calHost="hostname".}

\item{port}{Connection port. By default searches the .Renviron file for parameter=value pair of form calPort="1234".}

\item{username}{Username. By default searches the .Renviron file for parameter=value pair of form calUser="username". Refrain from using username with write privileges unless you absolutely know what you're doing.}

\item{password}{Password. By default searches the .Renviron file for parameter=value pair of form calPass="password".}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function exists to facilitate connecting to the calibrations database.
}
