% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{db_create}
\alias{db_create}
\title{Initial calibration database creation}
\usage{
db_create(path, overwrite = FALSE)
}
\arguments{
\item{path}{The path to the calibration SQLite database or the location where it should be created, with extension. Refer to calConnect for the default path refered to in that function as it may be appropriate to use here.}

\item{overwrite}{TRUE overwrites the database, if one exists in the same path.}
}
\value{
An SQLite database in the folder location specified by 'path'.
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Creates an SQLite database or replaces an existing database designed to hold instrument calibration data.
}
